fix tool naming
